refactor: split marketing optimization quality surfaces
